+++ routes/solids.js
@@ -1,23 +1,20 @@
 const express = require("express");
 const router = express.Router();
 const solids = require("../data/solids");
-let { ObjectID } = require('mongodb');
+const users = require("../data/users");
+const comments = require("../data/comments");
+const uuid = require('uuid');
 
 function isValidString(string) {
     if (!string) return false;
     if (typeof(string) !== 'string') return false;
-    if (string.trim().length === 0) return false;
+    if (string.trim() === '') return false;
     return true;
 }
 
 function isValidId(string) {
-    if (!isValidString) return false;
-    try {
-        const parsedId = ObjectId(string);
-        return true;
-    } catch (e) {
-        return false;
-    }
+    if (!isValidString(string)) return false;
+    return uuid.validate(string)
 }
 
 router.get("/:id", async (req, res) => {
@@ -27,7 +24,16 @@ router.get("/:id", async (req, res) => {
     }
     try {
         const solid = await solids.getSolidById(req.params.id);
-        res.status(200).json(result);
+        const creator = await users.getUserById(solid.postedBy);
+        let comments = [];
+        for (comment of solid.comments) {
+            comments.push(await comments.getCommentById(comment));
+        }
+        res.status(200).render("solids/solid.handlebars", {
+            solid: solid,
+            creator: creator,
+            comments: comments
+        });
     } catch (e) {
         if (e === 'solid not found') {
             res.status(404).json({error: "Solid not found"});
